% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageDescription.R
\docType{package}
\name{robFitConGraph-package}
\alias{robFitConGraph-package}
\alias{_PACKAGE}
\title{Robust estimation of scatter matrices subject to zero-constraints in their inverse.}
\description{
The package contains a function named after itself, \code{\link{robFitConGraph}}, 
two auxiliary functions \code{\link{find_eta}} and \code{\link{find_sigma1}} and a data set 
\code{\link{anxieties}}.
}
\details{
The function \code{robFitConGraph} provides two types of robust t-M-estimators of scatter 
subject to zero-constraints in the inverse. See the \code{robFitConGraph} function documentation 
for further details.

A tutorial including background information can be found on arxiv.org <doi:10.48550/arXiv.2204.04291>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://arxiv.org/abs/2204.04291}
}

}
