% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageDescription.R
\docType{package}
\name{robFitConGraph-package}
\alias{robFitConGraph-package}
\alias{_PACKAGE}
\title{A package for robust estimation of scatter matrices subject to zero- constraints in its inverse.}
\description{
robFitConGraph contains one function named after itself: \code{\link[robFitConGraph]{robFitConGraph}}.
}
